requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
